finapp-gamificado/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.32
sqlmodel==0.0.22
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.2.0
